+++ cmd/mimixbox/main.go
@@ -52,7 +52,7 @@ type options struct {
 
 var osExit = os.Exit
 
-const version = "0.24.1"
+const version = "0.25.0"
 
 const (
 	ExitSuccess int = iota // 0