+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 # Changelog
 All notable changes to this project will be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
-## [0.24.1] - 2021-11-27
+## [0.25.0] - 2021-11-27
 ### Added
  - hostid commad(Does not work properly)
+ - md5sum command.
  - seq command.
+ - sha1sum／sha256sum／sha512sum command
 ### Changed
 - error output destination of MimixBox from STDOUT to STDERR
 ## [0.22.0] - 2021-11-25